refactor dataloader code
