--- dynamic_graph_service/dataloader/.gitignore
@@ -1,2 +1,0 @@
-deps/
-data/
