+++ dynamic_graph_service/dataloader/apps/file_loader/group_producer.h
@@ -23,6 +23,8 @@ namespace dgs {
 namespace dataloader {
 namespace file {
 
+extern uint32_t batch_size;
+
 class GroupProducer {
 public:
   GroupProducer();
