+++ dynamic_graph_service/dataloader/include/dataloader/batch_builder.h
@@ -79,66 +79,6 @@ protected:
   std::vector<flatbuffers::Offset<RecordRep>> records_;
 };
 
-inline
-void BatchBuilder::AddVertexUpdate(VertexType vtype, VertexId vid, const std::vector<AttrInfo>& attrs) {
-  auto flat_attrs = AddAttributes(attrs);
-  auto flat_vertex = CreateVertexRecordRep(
-      builder_, vtype, vid, flat_attrs);
-  auto flat_record = CreateRecordRep(
-      builder_, RecordUnionRep_VertexRecordRep, flat_vertex.Union());
-  records_.emplace_back(flat_record);
-}
-
-inline
-void BatchBuilder::AddEdgeUpdate(EdgeType etype, VertexType src_vtype, VertexType dst_vtype,
-                                 VertexId src_vid, VertexId dst_vid, const std::vector<AttrInfo>& attrs) {
-  auto flat_attrs = AddAttributes(attrs);
-  auto flat_edge = CreateEdgeRecordRep(
-      builder_, etype, src_vtype, dst_vtype, src_vid, dst_vid, flat_attrs);
-  auto flat_record = CreateRecordRep(
-      builder_, RecordUnionRep_EdgeRecordRep, flat_edge.Union());
-  records_.emplace_back(flat_record);
-}
-
-inline
-flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<AttributeRecordRep>>>
-BatchBuilder::AddAttributes(const std::vector<AttrInfo>& attrs) {
-  std::vector<flatbuffers::Offset<AttributeRecordRep>> flat_attrs_vec;
-  flat_attrs_vec.reserve(attrs.size());
-  for (auto& a: attrs) {
-    auto flat_value_type = static_cast<AttributeValueTypeRep>(a.value_type);
-    auto flat_value_bytes = builder_.CreateVector(
-        reinterpret_cast<const int8_t*>(a.value_bytes.data()), a.value_bytes.size());
-    auto flat_attr = CreateAttributeRecordRep(
-        builder_, a.attr_type, flat_value_type, flat_value_bytes);
-    flat_attrs_vec.push_back(flat_attr);
-  }
-  return builder_.CreateVectorOfSortedTables(&flat_attrs_vec);
-}
-
-inline
-const uint8_t* BatchBuilder::GetBufferPointer() const {
-  return builder_.GetBufferPointer();
-}
-
-inline
-uint32_t BatchBuilder::GetBufferSize() const {
-  return builder_.GetSize();
-}
-
-inline
-void BatchBuilder::Finish() {
-  auto flat_records = builder_.CreateVector(records_);
-  auto batch = CreateRecordBatchRep(builder_, flat_records, partition_id_);
-  builder_.Finish(batch);
-}
-
-inline
-void BatchBuilder::Clear() {
-  builder_.Clear();
-  records_.clear();
-}
-
 }  // namespace dataloader
 }  // namespace dgs
