--- dynamic_graph_service/dataloader/include/dataloader/logging.h
@@ -1,34 +1,0 @@
-/* Copyright 2021 Alibaba Group Holding Limited. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-
-#ifndef DATALOADER_LOGGING_H_
-#define DATALOADER_LOGGING_H_
-
-#include <cstddef>
-#include <cstdint>
-#include <string>
-
-#include "glog/logging.h"
-
-namespace dgs {
-namespace dataloader {
-
-void InitGoogleLogging();
-void UninitGoogleLogging();
-
-}  // namespace dataloader
-}  // namespace dgs
-
-#endif // DATALOADER_LOGGING_H_
