--- dynamic_graph_service/dataloader/src/logging.cc
@@ -1,34 +1,0 @@
-/* Copyright 2021 Alibaba Group Holding Limited. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-
-#include "dataloader/logging.h"
-
-namespace dgs {
-namespace dataloader {
-
-void InitGoogleLogging() {
-  FLAGS_alsologtostderr = false;
-  FLAGS_colorlogtostderr = true;
-  FLAGS_log_dir = ".";
-  FLAGS_minloglevel = ::google::INFO;
-  ::google::InitGoogleLogging("dgs-dataloader");
-}
-
-void UninitGoogleLogging() {
-  ::google::ShutdownGoogleLogging();
-}
-
-}  // namespace dataloader
-}  // namespace dgs
